Move social card generation to page template
