+++ src/content/studio/config.json
@@ -1,6 +1,5 @@
 {
   "title": "Apollo Studio",
-  "subtitle": "Studio",
   "sidebar": {
     "Features": "/",
     "Get started": "/getting-started",
